Add files via upload
I use this beautiful project
https://github.com/LiheYoung/Depth-Anything
and this amazing project in combination
https://github.com/vikhyat/moondream

Get them both to run separately, then combine them into the depth project and use the code above and it should work for you like in the video. Use Mozilla over Edge as it is better for the camera setup. I use both USB camera and phone for the video